crawler/requirements.txt
praw==7.7.1
pandas==2.1.1
textblob==0.17.1
python-dotenv==1.0.0
fastapi==0.115.0
uvicorn==0.30.0
apscheduler==3.10.4
requests==2.31.0 